Cast to correct size ;)

+++ node_unix.go
@@ -26,7 +26,7 @@ func toStatT(i interface{}) (statT, bool) {
 func (s statUnix) dev() uint64   { return uint64(s.Dev) }
 func (s statUnix) ino() uint64   { return uint64(s.Ino) }
 func (s statUnix) nlink() uint64 { return uint64(s.Nlink) }
-func (s statUnix) uid() uint32   { return uint64(s.Uid) }
-func (s statUnix) gid() uint32   { return uint64(s.Gid) }
+func (s statUnix) uid() uint32   { return uint32(s.Uid) }
+func (s statUnix) gid() uint32   { return uint32(s.Gid) }
 func (s statUnix) rdev() uint64  { return uint64(s.Rdev) }
 func (s statUnix) size() int64   { return int64(s.Size) }
